Adding inital admin creation
When the root API URL is called it will check if there are any admins.
If not is will create one.
When the system is first set up, this can be used to create an initial admin account to them make proper accounts from.
Some more details in the handover doc.